pyproject.toml
[project]
name = "crawling-batch"
version = "0.1.0"
description = "Hot deal crawling batch service"
authors = [
    {name = "user1"}
]
requires-python = "^3.14"
dependencies = [
    "playwright (>=1.58.0,<2.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "apscheduler (>=3.11.2,<4.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "playwright-stealth (>=2.0.2,<3.0.0)"
]


[[tool.poetry.packages]]
include = "crawling_batch"
from = "src"

[project.scripts]
crawling-batch = "crawling_batch.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
